src/second_brain_online/pyproject.toml
[project]
name = "second-brain-course-online-pipeline"
version = "0.1.0"
description = "Self-paced course on production LLMs and RAG by teaching you how to build an AI assistant on top of your second brain."
readme = "ONLINE_README.md"
requires-python = ">=3.12.8"
authors = [
    {name = "Paul Iusztin", email = "p.b.iusztin@gmail.com"},
    {name = "Ernesto Larios", email = "elsoloscuro@hotmail.com"},
]
license = { text = "MIT" }

dependencies = [
    "click>=8.1.3",
    "datasets>=2.14.4",
    "loguru>=0.7.3",
    "pydantic>=2.8.2",
    "pydantic-settings>=2.7.0",
    "pymongo>=4.10.1",
    "boto3>=1.36.0",
    "langchain>=0.3.14",
    "langchain-mongodb>=0.4.0",
    "langchain-openai>=0.3.1",
    "langchain-core>=0.3.30",
    "ipykernel>=6.29.5",
    "gradio>=5.12.0",
    "smolagents>=1.4.1",
    "opentelemetry-api>=1.21.0",
    "opentelemetry-sdk>=1.21.0",
    "opentelemetry-instrumentation-requests>=0.42b0",
    "tiktoken>=0.5.2",
    "transformers>=4.36.0",
    "torch>=2.2.0",
    "opik>=0.1.0",
    "python-dotenv>=1.0.1",
    "pandas>=2.2.3",
    "langchain-community>=0.3.14",
    "sentence-transformers>=3.3.1",
    "rank-bm25>=0.2.2",
    "python-docx>=1.1.2",
    "PyPDF2>=3.0.1",
    "beautifulsoup4>=4.12.3",
    "unstructured>=0.16.13",
    "groq>=0.15.0",
    "comet_ml>=3.47.6",
]

[dependency-groups]
dev = [
    "pytest>=8.3.4",
    "ruff>=0.7.2",
]

[tool.pip]
extra-index-url = "https://download.pytorch.org/whl/cpu/torch_stable.html"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["second_brain_online"]

[tool.ruff]
target-version = "py312"